Changed some functionalities in backend

+++ 2april_backend.zip_expanded/EmailIndemintyMaintaince/src/main/java/com/axis/repository/VerifiedIndemnityRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import com.axis.model.VerifiedIndemnity;
 
 @Repository
-public interface VerifiedIndemnityRepository extends JpaRepository<VerifiedIndemnity, Long>{
+public interface VerifiedIndemnityRepository extends JpaRepository<VerifiedIndemnity, Integer>{
 	
 //	@Query("SELECT a  FROM VERIFIED_DETAILS WHERE a.accountNo = :accountNo")
 //	List<VerifiedIndemnity> findByAccountNo(String accountNo);
